requirements.txt
playwright==1.42.2
behave==1.2.7
behave-html-formatter==0.1.0
python-dotenv==1.0.0
